Add capability for UTF-16 character output
When set to Unicode, the output conversion function now can handle
characters outside the Basic Multilingual Plane, using UTF-16.
The output conversion function will detect whether a value passed in
is a low surrogate value and save it in a static variable.
If the next character is a correct high surrogate, the function will
return the correct unicode character, otherwise it will discard the
input.