Fully implement UTF-16BE
This patch adds full support for Unicode in the UTF-16BE (big-endian)
encoding. It ignores Byte-Order Markers and incorrect surrogate
sequences.

It also adds a Unicode program to the examples list, which prints out
"Hello World" and a couple of emojis outside the Basic Multilingual
Plane and a copyright sign as an example of a character inside the BMP.